IntervalEnum refactor, updated unit tests, getQuote with date range
