+++ src/main/java/edu/sdonohue/advancedjava/stocks/StockQuote.java
@@ -1,4 +1,4 @@
-package edu.sdonohue.advancedjava;
+package edu.sdonohue.advancedjava.stocks;
