+++ src/main/java/edu/sdonohue/advancedjava/stocks/StockServiceFactory.java
@@ -1,5 +1,7 @@
-package edu.sdonohue.advancedjava;
+package edu.sdonohue.advancedjava.stocks;
 
+
+import edu.sdonohue.advancedjava.database.DatabaseStockService;
 
 import javax.validation.constraints.NotNull;
