+++ src/test/java/edu/sdonohue/advancedjava/StockServiceFactoryTest.java
@@ -1,5 +1,7 @@
 package edu.sdonohue.advancedjava;
 
+import edu.sdonohue.advancedjava.stocks.StockService;
+import edu.sdonohue.advancedjava.stocks.StockServiceFactory;
 import org.junit.Test;
 
 import static org.junit.Assert.assertNotNull;
